Refactor EFCoreShellCommandHandler for cleaner and more efficient structure
`knownEfCommands()` function has been introduced to define the dotnet ef commands while bool check `matches()` now does not only check for command match but also checks for the commands existence in the known commands.

Lastly, added logic to execute input command. If the command is 'dotnet ef', an action popup is shown. If the command is different, it gets executed as before. Removed some redundant logging for cleaner code.