implemented TerminalShellCommandHandler (#188)
* Add command handling for shell terminal in EFCore plugin

Implemented a `TerminalShellCommandHandler` for the EFCore plugin to allow for executing EFCore commands from the terminal. Added new Kotlin class `EFCoreShellCommandHandler.kt` which contains the logic for matching and executing the given commands.

The `matches()` method checks if the input command matches with the known EFCore commands. If the command is 'dotnet ef', it triggers a quick actions popup in the terminal. If the command matches another known EFCore command, the corresponding action is executed.

Also updated the plugin's dependencies in `plugin.xml` and `build.gradle.kts` to include the terminal as a dependency for the EFCore plugin.

* from function to companion object

Co-authored-by: Andrii Rublov <seclerp.dev@gmail.com>

* Refactor EFCoreShellCommandHandler for better command parsing

Refactored EFCoreShellCommandHandler in the EFCore plugin to enhance the processing of entered shell commands. The string "dotnet ef" has been abstracted as a constant for better maintainability and consistency. Furthermore, the manner in which the command string is parsed and compared with known commands has been modified for more efficient execution.

---------

Co-authored-by: Alessio De Vito <alessio.devito@jetbrains.com>
Co-authored-by: Andrii Rublov <seclerp.dev@gmail.com>